Correctly handle C++ struct return values that contain capabilities.
The MIPS ABI passes everything from a superclass in integer registers (including floating point, where it makes no sense to do so and is likely an ABI written by looking at what a buggy gcc produces).  We only diverge from this in the case of capabilities, which must be passed in registers.